Minor updates to the feature that make it very ready to be shipped
Problems for future me: The text on the add CSV card does not match the seach card.

I broke a Library.text.tsx that fixed

I added a loading page for the CSV with live updates to the books that are being added. I also made the card able to be clicked anywhere for the upload, namely the image.

Big change was having the page now refresh when the books are added, causing them to actually show up right away